agentHost: make file paths in tool invocation messages nicer
Does proper markdown linking
