+++ src/vs/platform/agentHost/common/agentService.ts
@@ -226,7 +226,7 @@ export interface IAgentToolStartEvent extends IAgentProgressEventBase {
 	/** Human-readable display name for this tool. */
 	readonly displayName: string;
 	/** Message describing the tool invocation in progress (e.g., "Running `echo hello`"). */
-	readonly invocationMessage: string;
+	readonly invocationMessage: StringOrMarkdown;
 	/** A representative input string for display in the UI (e.g., the shell command). */
 	readonly toolInput?: string;
 	/** Hint for the renderer about how to display this tool (e.g., 'terminal' for shell commands, 'subagent' for subagent-spawning tools). */
