+++ src/vs/platform/agentHost/node/copilot/copilotToolDisplay.ts
@@ -3,8 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import type { PermissionRequest } from '@github/copilot-sdk';
 import { hasKey } from '../../../../base/common/types.js';
+import { URI } from '../../../../base/common/uri.js';
 import { localize } from '../../../../nls.js';
+import type { IAgentToolReadyEvent } from '../../common/agentService.js';
+import { StringOrMarkdown } from '../../common/state/protocol/state.js';
 
 // =============================================================================
 // Copilot CLI built-in tool interfaces
@@ -151,6 +155,22 @@ function truncate(text: string, maxLength: number): string {
 	return text.length > maxLength ? text.substring(0, maxLength - 3) + '...' : text;
 }
 
+/**
+ * Formats a file path as a markdown link `[](file-uri)` so it renders
+ * as a clickable file widget in the chat UI.
+ */
+function formatPathAsMarkdownLink(path: string): string {
+	return `[](${URI.file(path).toString()})`;
+}
+
+/**
+ * Wraps a localized message containing a markdown file link into a
+ * `StringOrMarkdown` object so the renderer treats it as markdown.
+ */
+function md(value: string): StringOrMarkdown {
+	return { markdown: value };
+}
+
 export function getToolDisplayName(toolName: string): string {
 	switch (toolName) {
 		case CopilotToolName.Bash: return localize('toolName.bash', "Bash");
@@ -176,7 +196,7 @@ export function getToolDisplayName(toolName: string): string {
 	}
 }
 
-export function getInvocationMessage(toolName: string, displayName: string, parameters: Record<string, unknown> | undefined): string {
+export function getInvocationMessage(toolName: string, displayName: string, parameters: Record<string, unknown> | undefined): StringOrMarkdown {
 	if (SHELL_TOOL_NAMES.has(toolName)) {
 		const args = parameters as ICopilotShellToolArgs | undefined;
 		if (args?.command) {
@@ -190,21 +210,21 @@ export function getInvocationMessage(toolName: string, displayName: string, para
 		case CopilotToolName.View: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolInvoke.viewFile', "Reading {0}", args.path);
+				return md(localize('toolInvoke.viewFile', "Reading {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolInvoke.view', "Reading file");
 		}
 		case CopilotToolName.Edit: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolInvoke.editFile', "Editing {0}", args.path);
+				return md(localize('toolInvoke.editFile', "Editing {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolInvoke.edit', "Editing file");
 		}
 		case CopilotToolName.Create: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolInvoke.createFile', "Creating {0}", args.path);
+				return md(localize('toolInvoke.createFile', "Creating {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolInvoke.create', "Creating file");
 		}
@@ -227,7 +247,7 @@ export function getInvocationMessage(toolName: string, displayName: string, para
 	}
 }
 
-export function getPastTenseMessage(toolName: string, displayName: string, parameters: Record<string, unknown> | undefined, success: boolean): string {
+export function getPastTenseMessage(toolName: string, displayName: string, parameters: Record<string, unknown> | undefined, success: boolean): StringOrMarkdown {
 	if (!success) {
 		return localize('toolComplete.failed', "\"{0}\" failed", displayName);
 	}
@@ -245,21 +265,21 @@ export function getPastTenseMessage(toolName: string, displayName: string, param
 		case CopilotToolName.View: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolComplete.viewFile', "Read {0}", args.path);
+				return md(localize('toolComplete.viewFile', "Read {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolComplete.view', "Read file");
 		}
 		case CopilotToolName.Edit: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolComplete.editFile', "Edited {0}", args.path);
+				return md(localize('toolComplete.editFile', "Edited {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolComplete.edit', "Edited file");
 		}
 		case CopilotToolName.Create: {
 			const args = parameters as ICopilotFileToolArgs | undefined;
 			if (args?.path) {
-				return localize('toolComplete.createFile', "Created {0}", args.path);
+				return md(localize('toolComplete.createFile', "Created {0}", formatPathAsMarkdownLink(args.path)));
 			}
 			return localize('toolComplete.create', "Created file");
 		}
@@ -341,5 +361,135 @@ export function getShellLanguage(toolName: string): string {
 	switch (toolName) {
 		case CopilotToolName.PowerShell: return 'powershell';
 		default: return 'shellscript';
+	}
+}
+
+// =============================================================================
+// Permission display
+//
+// Derives display fields from SDK permission requests for the tool
+// confirmation UI. Colocated with the tool-start display helpers above so
+// that formatting utilities (formatPathAsMarkdownLink, md, etc.) are shared.
+// =============================================================================
+
+export function tryStringify(value: unknown): string | undefined {
+	try {
+		return JSON.stringify(value);
+	} catch {
+		return undefined;
+	}
+}
+
+/**
+ * Extends the SDK's {@link PermissionRequest} with the known extra properties
+ * that arrive on the index-signature. The SDK defines these as `[key: string]: unknown`
+ * so this interface adds proper types for the fields we actually use.
+ */
+export interface ITypedPermissionRequest extends PermissionRequest {
+	/** File path — set for `read` permission requests. */
+	path?: string;
+	/** File path — set for `write` permission requests. */
+	fileName?: string;
+	/** Full shell command text — set for `shell` permission requests. */
+	fullCommandText?: string;
+	/** Human-readable intention describing the operation. */
+	intention?: string;
+	/** MCP server name — set for `mcp` permission requests. */
+	serverName?: string;
+	/** Tool name — set for `mcp` and `custom-tool` permission requests. */
+	toolName?: string;
+	/** Tool arguments — set for `custom-tool` permission requests. */
+	args?: Record<string, unknown>;
+}
+
+/** Safely extract a string value from an SDK field that may be `unknown` at runtime. */
+function str(value: unknown): string | undefined {
+	return typeof value === 'string' ? value : undefined;
+}
+
+/**
+ * Derives display fields from a permission request for the tool confirmation UI.
+ */
+export function getPermissionDisplay(request: ITypedPermissionRequest): {
+	confirmationTitle: string;
+	invocationMessage: StringOrMarkdown;
+	toolInput?: string;
+	/** Normalized permission kind for auto-approval routing. */
+	permissionKind: IAgentToolReadyEvent['permissionKind'];
+	/** File path extracted from the request. */
+	permissionPath?: string;
+} {
+	const path = str(request.path) ?? str(request.fileName);
+	const fullCommandText = str(request.fullCommandText);
+	const intention = str(request.intention);
+	const serverName = str(request.serverName);
+	const toolName = str(request.toolName);
+
+	switch (request.kind) {
+		case 'shell':
+			return {
+				confirmationTitle: localize('copilot.permission.shell.title', "Run in terminal"),
+				invocationMessage: intention ?? getInvocationMessage(CopilotToolName.Bash, getToolDisplayName(CopilotToolName.Bash), fullCommandText ? { command: fullCommandText } : undefined),
+				toolInput: fullCommandText,
+				permissionKind: 'shell',
+				permissionPath: path,
+			};
+		case 'custom-tool': {
+			// Custom tool overrides (e.g. our shell tool). Extract the actual
+			// tool args from the SDK's wrapper envelope.
+			const args = typeof request.args === 'object' && request.args !== null ? request.args as Record<string, unknown> : undefined;
+			const command = typeof args?.command === 'string' ? args.command : undefined;
+			const sdkToolName = str(request.toolName);
+			if (command && sdkToolName && isShellTool(sdkToolName)) {
+				return {
+					confirmationTitle: localize('copilot.permission.shell.title', "Run in terminal"),
+					invocationMessage: getInvocationMessage(sdkToolName, getToolDisplayName(sdkToolName), { command }),
+					toolInput: command,
+					permissionKind: 'shell',
+					permissionPath: path,
+				};
+			}
+			return {
+				confirmationTitle: toolName ?? localize('copilot.permission.default.title', "Permission request"),
+				invocationMessage: localize('copilot.permission.default.message', "Permission request"),
+				toolInput: args ? tryStringify(args) : tryStringify(request),
+				permissionKind: request.kind,
+				permissionPath: path,
+			};
+		}
+		case 'write':
+			return {
+				confirmationTitle: localize('copilot.permission.write.title', "Write file"),
+				invocationMessage: getInvocationMessage(CopilotToolName.Edit, getToolDisplayName(CopilotToolName.Edit), path ? { path } : undefined),
+				toolInput: tryStringify(path ? { path } : request) ?? undefined,
+				permissionKind: 'write',
+				permissionPath: path,
+			};
+		case 'mcp': {
+			const title = toolName ?? localize('copilot.permission.mcp.defaultTool', "MCP Tool");
+			return {
+				confirmationTitle: serverName ? `${serverName}: ${title}` : title,
+				invocationMessage: serverName ? `${serverName}: ${title}` : title,
+				toolInput: tryStringify({ serverName, toolName }) ?? undefined,
+				permissionKind: 'mcp',
+				permissionPath: path,
+			};
+		}
+		case 'read':
+			return {
+				confirmationTitle: localize('copilot.permission.read.title', "Read file"),
+				invocationMessage: intention ?? getInvocationMessage(CopilotToolName.View, getToolDisplayName(CopilotToolName.View), path ? { path } : undefined),
+				toolInput: tryStringify(path ? { path, intention } : request) ?? undefined,
+				permissionKind: 'read',
+				permissionPath: path,
+			};
+		default:
+			return {
+				confirmationTitle: localize('copilot.permission.default.title', "Permission request"),
+				invocationMessage: localize('copilot.permission.default.message', "Permission request"),
+				toolInput: tryStringify(request) ?? undefined,
+				permissionKind: request.kind,
+				permissionPath: path,
+			};
 	}
 }
